backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
gunicorn==21.2.0
anthropic>=0.34.0
pymupdf
opencv-python
numpy
python-dotenv
boto3 